WL#11825: Add Chinese collation for utf8mb4
Post-push fix for broken 32bit build:
strings/uca9-dump.cc:714:45: error: format %lu expects argument of type long unsigned int, but argument 3 has type std::map<int, int>::size_type {aka unsigned int} [-Werror=format=]

+++ strings/uca9-dump.cc
@@ -711,7 +711,7 @@ int dump_zh_hans(MY_UCA *uca, int *pageloaded, FILE *infile, FILE *outfile) {
   }
   fprintf(outfile, "\n};\n\n");
   fprintf(outfile, "extern const int ZH_HAN_WEIGHT_PAIRS = %lu;\n",
-          zh_han_to_single_weight_map.size());
+          static_cast<unsigned long>(zh_han_to_single_weight_map.size()));
 
   return 0;
 }
